Org: Move character.carp to top-level
Carp supports fetching libraries using git. In order to do so, it needs
to be certain of which file to load as the main file in the library. By
default it looks for main.carp, but some libraries use the repo name.
This move checks if simply moving character.carp will get the git
based loading to work.