Character: Make the character type abstract
I've refactored this tiny Character type to serve as a wrapper ADT
around fixed character sets. This makes them a little harder to work
with in general, but gives us some power. If users define a minimal
set of fixed characters as an enum and an interface comprised of the
small functions:

`char`: Converts an enum member value to a Char value.
`character`: Converts a Char value to an enum set member.
(These are just projection functions)

we're able to make transformations between differently defined sets, and
get validation over the character sets for free.